Add: export of mt.condition and mt.semaphore API; Update: JSON example; Minor Fix: portability issue of printf() with %zu; Minor Add: indenting dots for multiline interactive mode.

+++ kernel/daoArray.c
@@ -27,7 +27,7 @@ void* dao_malloc( size_t size )
 {
 	void *p = malloc( size );
 	if( size && p == NULL ){
-		printf( "memory allocation %zu failed\n", size );
+		printf( "memory allocation %lu failed\n", (unsigned long)size );
 		exit(1);
 	}
 	return p;
@@ -36,7 +36,7 @@ void* dao_calloc( size_t nmemb, size_t size )
 {
 	void *p = calloc( nmemb, size );
 	if( size && p == NULL ){
-		printf( "memory allocation %zu failed\n", size * nmemb );
+		printf( "memory allocation %lu failed\n", (unsigned long)( size * nmemb ) );
 		exit(1);
 	}
 	return p;
@@ -45,7 +45,7 @@ void* dao_realloc( void *ptr, size_t size )
 {
 	void *p = realloc( ptr, size );
 	if( size && p == NULL ){
-		printf( "memory allocation %zu failed\n", size );
+		printf( "memory allocation %lu failed\n", (unsigned long)size );
 		exit(1);
 	}
 	return p;
